begin addressing PR comments
